requirements.txt
# Core dependencies
gradio>=4.16.0,<5.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
numpy>=1.24.0,<2.0.0

# PyTorch and audio processing
torch>=2.0.0,<3.0.0
torchaudio>=2.0.0,<3.0.0
transformers>=4.35.0,<5.0.0
pyannote.audio>=3.0.0,<4.0.0

# AWS and API dependencies
boto3>=1.28.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# Machine learning
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0

# Utility dependencies
tenacity>=8.2.0,<9.0.0
tqdm>=4.65.0,<5.0.0
openpyxl>=3.1.0,<4.0.0

# Data processing
scipy>=1.10.0,<2.0.0

# System monitoring
psutil>=5.9.0,<6.0.0

# Optional: for better performance
accelerate>=0.25.0,<1.0.0 

huggingface-hub>=0.25.0,<1.0.0